requirements.txt
pandas
matplotlib
numpy
opencv-python==4.5.5.62
streamlit
Pillow
opencv-contrib-python